Initial hack to make thorium event driven, instead of polling.
Thorium is similar in most ways to the rest of salt- with one major distringuishing factor-- we are dealing with state over time, meaning the calls to these module methods aren't a one-time deal-- we actually want them to build and maintain state.

With that in mind, the goal in this hack is to have the modules create instances of themselves-- registering for whatever events that they require (instead of polling forever). I've only made this work for the example up on docs.saltstack.com. As you can see in reg and file-- we will create coroutines on the IOLoop waiting on the various events that we require. This way we don't have to pass all events to every module-- and we still only have to consume the events once.

At this point there are a few outstanding questions:
	- what to do when SLS files change (restart the process? Un-register?)
	- should these fire events when they change (they are now)-- or should these events be internal to thorium only
	- it seems like the module interface could use some work, as each one ends up implementing a similar while true and yield block. Maybe it'd be better to do some sort of callback interface?
	- do we want to use the regular states interface? If so, how do we deal with requisites etc. What I've done here is just use the __low__ which normalizes all requisites together-- and I'm using that as "you want to be called when one of those changes"-- which in a callback mechanism might actually be fairly clean.